ez_monitor/requirements.txt
Flask==2.1.0
psutil==5.9.0
flask-cors==5.0.0
py-cpuinfo==9.0.0
GPUtil==1.4.0